requirements/base.txt
# Requeriments base.

Django==5.0
django-cors-headers==4.3.1
django-filter==21.1
djangorestframework==3.14.0
django-environ==0.11.2
djangorestframework-jsonapi==6.1.0

#recaptcha
django-recaptcha==4.0.0

#oauth toolkit
django-oauth-toolkit==2.3.0


# database
psycopg2==2.9.9
psycopg2-binary==2.9.9